Clean up PiGUI Lua
data/pigui/pigui.lua is now contained in several modules:
  - data/pigui/baseui.lua - a base set of the ui functions (forwarded + wrappers + basic shared config)
  - data/pigui/libs/forwarded.lua - functions directly forwarded from C++
  - data/pigui/libs/wrappers.lua - convenience wrappers (mostly)
  - data/pigui/libs/buttons.lua - button convenience functions
  - data/pigui/libs/icons.lua - icon related stuff
  - data/pigui/libs/text.lua - text display related
  - data/pigui/libs/gauge.lua - gauge convenience functions
  - data/pigui/libs/radial-menu.lua - radial menu convenience functions

Add documentation to UI functions implemented in Lua

Import begone!

Spaces vs tabs and indentation cleanup

Usage of PiGUI should still be done through `local ui = require 'pigui'`
