+++ data/pigui/libs/modal-win.lua
@@ -1,7 +1,7 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local ui = import 'pigui/pigui.lua'
+local ui = require 'pigui'
 
 local modals = {}
 local modalStack = {}
