+++ data/pigui/modules/station-view/03-commodityMarket.lua
@@ -1,11 +1,11 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local Lang = import 'Lang'
-local StationView = import 'pigui/views/station-view'
-local CommodityWidget = import 'pigui/libs/commodity-market.lua'
+local Lang = require 'Lang'
+local StationView = require 'pigui.views.station-view'
+local CommodityWidget = require 'pigui.libs.commodity-market'
 
-local ui = import 'pigui/pigui.lua'
+local ui = require 'pigui'
 local pionillium = ui.fonts.pionillium
 local orbiteer = ui.fonts.orbiteer
 local l = Lang.GetResource("ui-core")
