+++ data/pigui/modules/station-view/04-shipMarket.lua
@@ -1,17 +1,17 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local Lang = import 'Lang'
-local Game = import 'Game'
-local Format = import 'Format'
-local ShipDef = import 'ShipDef'
-local Equipment = import 'Equipment'
-local StationView = import 'pigui/views/station-view'
-local Table = import 'pigui/libs/table.lua'
-local PiImage = import 'ui/PiImage'
-local ModelSpinner = import 'PiGui.Modules.ModelSpinner'
+local Lang = require 'Lang'
+local Game = require 'Game'
+local Format = require 'Format'
+local ShipDef = require 'ShipDef'
+local Equipment = require 'Equipment'
+local StationView = require 'pigui.views.station-view'
+local Table = require 'pigui.libs.table'
+local PiImage = require 'ui/PiImage'
+local ModelSpinner = require 'PiGui.Modules.ModelSpinner'
 
-local ui = import 'pigui/pigui.lua'
+local ui = require 'pigui'
 local pionillium = ui.fonts.pionillium
 local orbiteer = ui.fonts.orbiteer
 local l = Lang.GetResource("ui-core")
