+++ data/pigui/modules/station-view/06-shipRepairs.lua
@@ -1,16 +1,16 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local ui = import 'pigui/pigui.lua'
-local StationView = import 'pigui/views/station-view'
-local ShipDef = import("ShipDef")
-local Game = import "Game"
-local Rand = import "Rand"
-local InfoFace = import 'ui/PiguiFace'
-local Format = import "Format"
-local Character = import "Character"
-local ModalWindow = import 'pigui/libs/modal-win.lua'
-local Lang = import 'Lang'
+local ui = require 'pigui'
+local StationView = require 'pigui.views.station-view'
+local ShipDef = require "ShipDef"
+local Game = require "Game"
+local Rand = require "Rand"
+local InfoFace = require 'ui/PiguiFace'
+local Format = require "Format"
+local Character = require "Character"
+local ModalWindow = require 'pigui.libs.modal-win'
+local Lang = require 'Lang'
 local l = Lang.GetResource("ui-core")
 
 local rescaleVector = ui.rescaleUI(Vector2(1, 1), Vector2(1600, 900), true)
