+++ data/pigui/modules/station-view/07-police.lua
@@ -1,19 +1,19 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local Game = import "Game"
-local Rand = import "Rand"
-local ui = import 'pigui/pigui.lua'
-local StationView = import 'pigui/views/station-view'
-local Lang = import 'Lang'
-local Legal = import "Legal"
-local utils = import "utils"
-local InfoFace = import 'ui/PiguiFace'
-local Format = import "Format"
-local Character = import "Character"
+local Game = require "Game"
+local Rand = require "Rand"
+local ui = require 'pigui'
+local StationView = require 'pigui.views.station-view'
+local Lang = require 'Lang'
+local Legal = require "Legal"
+local utils = require "utils"
+local InfoFace = require 'ui/PiguiFace'
+local Format = require "Format"
+local Character = require "Character"
 local l = Lang.GetResource("ui-core")
 
-local ModalWindow = import 'pigui/libs/modal-win.lua'
+local ModalWindow = require 'pigui.libs.modal-win'
 
 local rescaleVector = ui.rescaleUI(Vector2(1, 1), Vector2(1600, 900), true)
