+++ data/pigui/views/info-view.lua
@@ -1,8 +1,8 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local ui = import 'pigui/pigui.lua'
-local TabView = import 'pigui/views/tab-view.lua'
+local ui = require 'pigui'
+local TabView = require 'pigui.views.tab-view'
 
 local infoView
